+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 - Switched to Scala.js-1.7.0.
+- `hasher` can be used now as `OutputStream`.
 
 ## [2.8.0] - 2021-07-16
 - Introduced `doneXor`.